Update README.md
SpamLess is the website which will have reviews of different categories. User will be able to read and write reviews. Every review written will go though Spam Detection Algorithm and it status will be updated as "Truthfull" if spam not detected and will be as "Deceptive" if spam will be detected.